+++ Sexy.TodLib/TodCommon.cpp
@@ -872,7 +872,7 @@ void TodDrawImageCenterScaledF(Graphics* g, Image* theImage, float thePosX, floa
 }
 
 //0x512AC0
-unsigned long AverageNearByPixels(MemoryImage* theImage, unsigned long* thePixel, int x, int y)
+uint32_t AverageNearByPixels(MemoryImage* theImage, uint32_t* thePixel, int x, int y)
 {
 	int aRed = 0;
 	int aGreen = 0;
@@ -890,7 +890,7 @@ unsigned long AverageNearByPixels(MemoryImage* theImage, unsigned long* thePixel
 		{
 			if ((x != 0 || j != -1) && (x != theImage->mWidth - 1 || j != 1) && (y != 0 || i != -1) && (y != theImage->mHeight - 1 || i != 1))
 			{
-				unsigned long aPixel = *(thePixel + i * theImage->mWidth + j);
+				uint32_t aPixel = *(thePixel + i * theImage->mWidth + j);
 				if (aPixel & 0xFF000000UL)  // 如果不是透明像素
 				{
 					aRed += (aPixel >> 16) & 0x000000FFUL;
@@ -928,7 +928,7 @@ void FixPixelsOnAlphaEdgeForBlending(Image* theImage)
 	PerfTimer aTimer;
 	aTimer.Start();
 
-	unsigned long* aBitsPtr = aImage->mBits;
+	uint32_t* aBitsPtr = aImage->mBits;
 	for (int y = 0; y < theImage->mHeight; y++)
 	{
 		for (int x = 0; x < theImage->mWidth; x++)
